+++ controllers/lrgs-config-controller/src/lrgs/mod.rs
@@ -1,0 +1,2 @@
+pub mod lrgs;
+pub mod password_file;
